+++ src/Convert/numpy.jl
@@ -99,6 +99,7 @@ const NUMPY_SIMPLE_TYPES = [
 
 function init_numpy()
     # simple numeric scalar types
+    priority = 1
     for (t, T) in NUMPY_SIMPLE_TYPES
         isbool = occursin("bool", t)
         isint = occursin("int", t) || isbool
@@ -112,49 +113,79 @@ function init_numpy()
         rule = pyconvert_rule_numpysimplevalue{T,false}()
         saferule = pyconvert_rule_numpysimplevalue{T,true}()
 
-        pyconvert_add_rule(name, T, saferule, PYCONVERT_PRIORITY_ARRAY)
-        isuint && pyconvert_add_rule(name, UInt, sizeof(T) ≤ sizeof(UInt) ? saferule : rule)
-        isuint && pyconvert_add_rule(name, Int, sizeof(T) < sizeof(Int) ? saferule : rule)
-        isint &&
-            !isuint &&
-            pyconvert_add_rule(name, Int, sizeof(T) ≤ sizeof(Int) ? saferule : rule)
-        isint && pyconvert_add_rule(name, Integer, rule)
-        isfloat && pyconvert_add_rule(name, Float64, saferule)
-        isreal && pyconvert_add_rule(name, Real, rule)
-        iscomplex && pyconvert_add_rule(name, ComplexF64, saferule)
-        iscomplex && pyconvert_add_rule(name, Complex, rule)
-        isnumber && pyconvert_add_rule(name, Number, rule)
+        if isnumber
+            pyconvert_add_rule(name, Number, Number, rule)
+            if isreal
+                pyconvert_add_rule(name, Real, Number, rule)
+                pyconvert_add_rule(name, AbstractFloat, Number, rule)
+            else
+                pyconvert_add_rule(name, Complex, Number, rule)
+            end
+            if isint
+                pyconvert_add_rule(name, Integer, Number, rule)
+                pyconvert_add_rule(name, isuint ? Unsigned : Signed, Number, rule)
+                if isuint
+                    pyconvert_add_rule(
+                        name,
+                        UInt,
+                        Number,
+                        sizeof(T) ≤ sizeof(UInt) ? saferule : rule,
+                    )
+                else
+                    pyconvert_add_rule(
+                        name,
+                        Int,
+                        Number,
+                        sizeof(T) ≤ sizeof(Int) ? saferule : rule,
+                    )
+                end
+            elseif isfloat
+                pyconvert_add_rule(name, Float64, Number, saferule)
+            elseif iscomplex
+                pyconvert_add_rule(name, ComplexF64, Number, saferule)
+            end
+            pyconvert_add_rule(name, T, Number, saferule)
+        end
     end
 
     # datetime64
-    pyconvert_add_rule(
-        "numpy:datetime64",
-        DateTime64,
-        pyconvert_rule_datetime64,
-        PYCONVERT_PRIORITY_ARRAY,
-    )
-    pyconvert_add_rule("numpy:datetime64", InlineDateTime64, pyconvert_rule_datetime64)
-    pyconvert_add_rule(
-        "numpy:datetime64",
+    pyconvert_add_rule("numpy:datetime64", InlineDateTime64, InlineDateTime64, pyconvert_rule_datetime64)
+    pyconvert_add_rule("numpy:datetime64",
         NumpyDates.DatesInstant,
-        pyconvert_rule_datetime64,
-    )
-    pyconvert_add_rule("numpy:datetime64", Missing, pyconvert_rule_datetime64)
-    pyconvert_add_rule("numpy:datetime64", Nothing, pyconvert_rule_datetime64)
+        NumpyDates.DatesInstant,
+        pyconvert_rule_datetime64)
+    pyconvert_add_rule("numpy:datetime64", Missing, Missing, pyconvert_rule_datetime64)
+    pyconvert_add_rule("numpy:datetime64", Nothing, Nothing, pyconvert_rule_datetime64)
 
     # timedelta64
-    pyconvert_add_rule(
+    pyconvert_add_rule("numpy:timedelta64", InlineTimeDelta64, InlineTimeDelta64, pyconvert_rule_timedelta64)
+    pyconvert_add_rule("numpy:timedelta64",
+        NumpyDates.DatesPeriod,
+        NumpyDates.DatesPeriod,
+        pyconvert_rule_timedelta64)
+    pyconvert_add_rule("numpy:timedelta64", Missing, Missing, pyconvert_rule_timedelta64)
+    pyconvert_add_rule("numpy:timedelta64", Nothing, Nothing, pyconvert_rule_timedelta64)
+    for (t, T) in NUMPY_SIMPLE_TYPES
+        pyconvert_add_rule_high_priority(
+            "numpy:$t",
+            T,
+            Any,
+            pyconvert_rule_numpysimplevalue{T,true}(),
+            priority,
+        )
+    end
+    pyconvert_add_rule_high_priority(
+        "numpy:datetime64",
+        DateTime64,
+        Any,
+        pyconvert_rule_datetime64,
+        priority,
+    )
+    pyconvert_add_rule_high_priority(
         "numpy:timedelta64",
         TimeDelta64,
+        Any,
         pyconvert_rule_timedelta64,
-        PYCONVERT_PRIORITY_ARRAY,
+        priority,
     )
-    pyconvert_add_rule("numpy:timedelta64", InlineTimeDelta64, pyconvert_rule_timedelta64)
-    pyconvert_add_rule(
-        "numpy:timedelta64",
-        NumpyDates.DatesPeriod,
-        pyconvert_rule_timedelta64,
-    )
-    pyconvert_add_rule("numpy:timedelta64", Missing, pyconvert_rule_timedelta64)
-    pyconvert_add_rule("numpy:timedelta64", Nothing, pyconvert_rule_timedelta64)
 end