+++ src/Convert/pandas.jl
@@ -2,11 +2,12 @@ pyconvert_rule_pandas_na(::Type{Nothing}, x::Py) = pyconvert_return(nothing)
 pyconvert_rule_pandas_na(::Type{Missing}, x::Py) = pyconvert_return(missing)
 
 function init_pandas()
-    pyconvert_add_rule(
-        "pandas._libs.missing:NAType",
+    pyconvert_add_rule_high_priority(
+        "pandas.api.typing:NAType",
         Missing,
+        Any,
         pyconvert_rule_pandas_na,
-        PYCONVERT_PRIORITY_CANONICAL,
+        0,
     )
-    pyconvert_add_rule("pandas._libs.missing:NAType", Nothing, pyconvert_rule_pandas_na)
+    pyconvert_add_rule("pandas.api.typing:NAType", Nothing, Nothing, pyconvert_rule_pandas_na)
 end