+++ src/Wrap/Wrap.jl
@@ -32,55 +32,77 @@ include("PyTable.jl")
 include("PyPandasDataFrame.jl")
 
 function __init__()
-    priority = PYCONVERT_PRIORITY_ARRAY
-    pyconvert_add_rule("<arraystruct>", PyArray, pyconvert_rule_array_nocopy, priority)
-    pyconvert_add_rule("<arrayinterface>", PyArray, pyconvert_rule_array_nocopy, priority)
-    pyconvert_add_rule("<array>", PyArray, pyconvert_rule_array_nocopy, priority)
-    pyconvert_add_rule("<buffer>", PyArray, pyconvert_rule_array_nocopy, priority)
-
-    priority = PYCONVERT_PRIORITY_CANONICAL
-    pyconvert_add_rule(
-        "collections.abc:Iterable",
-        PyIterable,
-        pyconvert_rule_iterable,
+    priority = 0
+    Convert.pyconvert_add_rule_high_priority(
+        "pandas:DataFrame",
+        PyPandasDataFrame,
+        Any,
+        pyconvert_rule_pandasdataframe,
         priority,
     )
-    pyconvert_add_rule(
-        "collections.abc:Sequence",
+    Convert.pyconvert_add_rule_high_priority(
+        "pandas.api.extensions:ExtensionArray",
         PyList,
+        Any,
         pyconvert_rule_sequence,
         priority,
     )
-    pyconvert_add_rule("collections.abc:Set", PySet, pyconvert_rule_set, priority)
-    pyconvert_add_rule("collections.abc:Mapping", PyDict, pyconvert_rule_mapping, priority)
-    pyconvert_add_rule("io:IOBase", PyIO, pyconvert_rule_io, priority)
-    pyconvert_add_rule("_io:_IOBase", PyIO, pyconvert_rule_io, priority)
-    for typename in ["pandas.core.frame:DataFrame", "pandas:DataFrame"]
-        pyconvert_add_rule(
-            typename,
-            PyPandasDataFrame,
-            pyconvert_rule_pandasdataframe,
-            priority,
-        )
-    end
-    for typename in ["pandas.core.arrays.base:ExtensionArray", "pandas.api.extensions:ExtensionArray"]
-        pyconvert_add_rule(
-            typename,
-            PyList,
-            pyconvert_rule_sequence,
-            priority,
-        )
-    end
+    Convert.pyconvert_add_rule_high_priority(
+        "collections.abc:Sequence",
+        PyList,
+        Any,
+        pyconvert_rule_sequence,
+        priority,
+    )
+    Convert.pyconvert_add_rule_high_priority(
+        "collections.abc:Set",
+        PySet,
+        Any,
+        pyconvert_rule_set,
+        priority,
+    )
+    Convert.pyconvert_add_rule_high_priority(
+        "collections.abc:Mapping",
+        PyDict,
+        Any,
+        pyconvert_rule_mapping,
+        priority,
+    )
+    Convert.pyconvert_add_rule_high_priority(
+        "collections.abc:Iterable",
+        PyIterable,
+        Any,
+        pyconvert_rule_iterable,
+        priority,
+    )
+    Convert.pyconvert_add_rule_high_priority(
+        "io:IOBase",
+        PyIO,
+        Any,
+        pyconvert_rule_io,
+        priority,
+    )
 
-    priority = PYCONVERT_PRIORITY_NORMAL
-    pyconvert_add_rule("<arraystruct>", Array, pyconvert_rule_array, priority)
-    pyconvert_add_rule("<arrayinterface>", Array, pyconvert_rule_array, priority)
-    pyconvert_add_rule("<array>", Array, pyconvert_rule_array, priority)
-    pyconvert_add_rule("<buffer>", Array, pyconvert_rule_array, priority)
-    pyconvert_add_rule("<arraystruct>", AbstractArray, pyconvert_rule_array, priority)
-    pyconvert_add_rule("<arrayinterface>", AbstractArray, pyconvert_rule_array, priority)
-    pyconvert_add_rule("<array>", AbstractArray, pyconvert_rule_array, priority)
-    pyconvert_add_rule("<buffer>", AbstractArray, pyconvert_rule_array, priority)
+    pyconvert_add_rule("<arraystruct>", AbstractArray, AbstractArray, pyconvert_rule_array)
+    pyconvert_add_rule("<arrayinterface>", AbstractArray, AbstractArray, pyconvert_rule_array)
+    pyconvert_add_rule("<array>", AbstractArray, AbstractArray, pyconvert_rule_array)
+    pyconvert_add_rule("<buffer>", AbstractArray, AbstractArray, pyconvert_rule_array)
+    Convert.pyconvert_add_rule_high_priority(
+        "<arraystruct>",
+        PyArray,
+        Any,
+        pyconvert_rule_array_nocopy,
+        1,
+    )
+    Convert.pyconvert_add_rule_high_priority(
+        "<arrayinterface>",
+        PyArray,
+        Any,
+        pyconvert_rule_array_nocopy,
+        1,
+    )
+    Convert.pyconvert_add_rule_high_priority("<array>", PyArray, Any, pyconvert_rule_array_nocopy, 1)
+    Convert.pyconvert_add_rule_high_priority("<buffer>", PyArray, Any, pyconvert_rule_array_nocopy, 1)
 end
 
 end